requirements.txt
bs4
requests
fake-useragent
fastapi
motor
langchain_openai
langchain
ruamel_yaml
dropbox
jinja2
pdfkit